BL-325 a bit of restructuring, logged initial signed request, but CORS error on upload to AWS
